third_party/blink/web_tests/external/wpt/webrtc/protocol/h264-profile-levels.https-expected.txt

This is a testharness.js-based test.
NOTRUN Level 1 H264 video is appropriately constrained Level 1 H264 video is not supported
FAIL Level 2 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 396 but got 2025
FAIL Level 3 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 1620 but got 2025
PASS Level 4 H264 video is appropriately constrained
PASS Level 5 H264 video is appropriately constrained
NOTRUN Level 6 H264 video is appropriately constrained Level 6 H264 video is not supported
NOTRUN Level 1.1 H264 video is appropriately constrained Level 1.1 H264 video is not supported
NOTRUN Level 1.2 H264 video is appropriately constrained Level 1.2 H264 video is not supported
NOTRUN Level 1.3 H264 video is appropriately constrained Level 1.3 H264 video is not supported
FAIL Level 2.1 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 792 but got 2025
FAIL Level 2.2 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 1620 but got 2025
PASS Level 3.1 H264 video is appropriately constrained
PASS Level 3.2 H264 video is appropriately constrained
PASS Level 4.1 H264 video is appropriately constrained
PASS Level 4.2 H264 video is appropriately constrained
PASS Level 5.1 H264 video is appropriately constrained
PASS Level 5.2 H264 video is appropriately constrained
NOTRUN Level 6.1 H264 video is appropriately constrained Level 6.1 H264 video is not supported
NOTRUN Level 6.2 H264 video is appropriately constrained Level 6.2 H264 video is not supported
Harness: the test ran to completion.
